requirements.txt
pyshark
pyserial
